@@ -106,7 +106,7 @@
     "moment": "^2.17.1",
     "mongoose": "^4.8.0",
     "morgan": "^1.7.0",
-    "node-sass": "^4.0.0",
+    "node-sass": "^4.5.0",
     "normalize.css": "^5.0.0",
     "passport": "^0.3.2",
     "passport-github2": "^0.1.10",
@@ -137,12 +137,15 @@
     "cheerio": "^0.22.0",
     "codecov": "^1.0.1",
     "enzyme": "^2.0.0",
-    "eslint": "^3.0.1",
+    "eslint": "^3.14.1",
+    "eslint-config-airbnb": "^14.0.0",
     "eslint-config-standard": "^6.0.0",
     "eslint-config-standard-react": "^4.0.0",
     "eslint-plugin-babel": "^4.0.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^3.0.2",
     "eslint-plugin-promise": "^3.0.0",
-    "eslint-plugin-react": "^6.0.0",
+    "eslint-plugin-react": "^6.9.0",
     "eslint-plugin-standard": "^2.0.0",
     "express": "^4.14.0",
     "karma": "^1.0.0",
