@@ -5,7 +5,7 @@ const webpackConfig = require('../config/webpack.config');
 const project = require('../config/project.config');
 
 // Wrapper around webpack to promisify its compiler and supply friendly logging
-const webpackCompiler = (webpackConfig) =>
+const webpackCompiler = webpackConfig =>
   new Promise((resolve, reject) => {
     const compiler = webpack(webpackConfig);
 
@@ -37,7 +37,7 @@ const compile = () => {
   debug('Starting compiler.');
   return Promise.resolve()
     .then(() => webpackCompiler(webpackConfig))
-    .then(stats => {
+    .then((stats) => {
       if (stats.warnings.length && project.compiler_fail_on_warning) {
         throw new Error('Config set to fail on warning, exiting with status code "1".');
       }
