@@ -8,22 +8,22 @@ module.exports = {
   // NOTE: In development, we use an explicit public path when the assets
   // are served webpack by to fix this issue:
   // http://stackoverflow.com/questions/34133808/webpack-ots-parsing-error-loading-fonts/34133809#34133809
-  development : (config) => ({
-    compiler_public_path : `http://${config.server_host}:${config.server_port}/`
+  development: config => ({
+    compiler_public_path: `http://${config.server_host}:${config.server_port}/`,
   }),
 
   // ======================================================
   // Overrides when NODE_ENV === 'production'
   // ======================================================
-  production : (config) => ({
-    compiler_public_path     : '/',
-    compiler_fail_on_warning : false,
-    compiler_hash_type       : 'chunkhash',
-    compiler_devtool         : null,
-    compiler_stats           : {
-      chunks       : true,
-      chunkModules : true,
-      colors       : true
-    }
-  })
+  production: config => ({
+    compiler_public_path: '/',
+    compiler_fail_on_warning: false,
+    compiler_hash_type: 'chunkhash',
+    compiler_devtool: null,
+    compiler_stats: {
+      chunks: true,
+      chunkModules: true,
+      colors: true,
+    },
+  }),
 };
