@@ -5,69 +5,69 @@ const debug = require('debug')('app:config:karma');
 
 debug('Creating configuration.');
 const karmaConfig = {
-  basePath : '../', // project root in relation to bin/karma.js
-  files    : [
+  basePath: '../', // project root in relation to bin/karma.js
+  files: [
     {
-      pattern  : `./${project.dir_test}/test-bundler.js`,
-      watched  : false,
-      served   : true,
-      included : true
-    }
+      pattern: `./${project.dir_test}/test-bundler.js`,
+      watched: false,
+      served: true,
+      included: true,
+    },
   ],
-  singleRun     : !argv.watch,
-  frameworks    : ['mocha'],
-  reporters     : ['mocha'],
-  preprocessors : {
-    [`${project.dir_test}/test-bundler.js`] : ['webpack']
+  singleRun: !argv.watch,
+  frameworks: ['mocha'],
+  reporters: ['mocha'],
+  preprocessors: {
+    [`${project.dir_test}/test-bundler.js`]: ['webpack'],
   },
-  browsers : ['PhantomJS'],
-  webpack  : {
-    devtool : 'cheap-module-source-map',
-    resolve : Object.assign({}, webpackConfig.resolve, {
-      alias : Object.assign({}, webpackConfig.resolve.alias, {
-        sinon : 'sinon/pkg/sinon.js'
-      })
+  browsers: ['PhantomJS'],
+  webpack: {
+    devtool: 'cheap-module-source-map',
+    resolve: Object.assign({}, webpackConfig.resolve, {
+      alias: Object.assign({}, webpackConfig.resolve.alias, {
+        sinon: 'sinon/pkg/sinon.js',
+      }),
     }),
-    plugins : webpackConfig.plugins,
-    module  : {
-      noParse : [
-        /\/sinon\.js/
+    plugins: webpackConfig.plugins,
+    module: {
+      noParse: [
+        /\/sinon\.js/,
       ],
-      loaders : webpackConfig.module.loaders.concat([
+      loaders: webpackConfig.module.loaders.concat([
         {
-          test   : /sinon(\\|\/)pkg(\\|\/)sinon\.js/,
-          loader : 'imports?define=>false,require=>false'
-        }
-      ])
+          test: /sinon(\\|\/)pkg(\\|\/)sinon\.js/,
+          loader: 'imports?define=>false,require=>false',
+        },
+      ]),
     },
     // Enzyme fix, see:
     // https://github.com/airbnb/enzyme/issues/47
-    externals : Object.assign({}, webpackConfig.externals, {
-      'react/addons'                   : true,
-      'react/lib/ExecutionEnvironment' : true,
-      'react/lib/ReactContext'         : 'window'
+    externals: Object.assign({}, webpackConfig.externals, {
+      'react/addons': true,
+      'react/lib/ExecutionEnvironment': true,
+      'react/lib/ReactContext': 'window',
     }),
-    sassLoader : webpackConfig.sassLoader
+    sassLoader: webpackConfig.sassLoader,
   },
-  webpackMiddleware : {
-    noInfo : true
+  webpackMiddleware: {
+    noInfo: true,
   },
-  coverageReporter : {
-    reporters : project.coverage_reporters
-  }
+  coverageReporter: {
+    reporters: project.coverage_reporters,
+  },
 };
 
 if (project.globals.__COVERAGE__) {
   karmaConfig.reporters.push('coverage');
   karmaConfig.webpack.module.preLoaders = [{
-    test    : /\.(js|jsx)$/,
-    include : new RegExp(project.dir_client),
-    exclude : /node_modules/,
-    loader  : 'babel',
-    query   : Object.assign({}, project.compiler_babel, {
-      plugins : (project.compiler_babel.plugins || []).concat('istanbul')
-    })
+    test: /\.(js|jsx)$/,
+    include: new RegExp(project.dir_client),
+    exclude: /node_modules/,
+    loader: 'babel',
+    query: Object.assign({}, project.compiler_babel, {
+      plugins: (project.compiler_babel.plugins || []).concat('istanbul'),
+    }),
   }];
 }
 
-module.exports = (cfg) => cfg.set(karmaConfig);
+module.exports = cfg => cfg.set(karmaConfig);
