@@ -33,7 +33,7 @@ const config = {
   compiler_babel : {
     cacheDirectory : true,
     plugins        : ['transform-runtime'],
-    presets        : ['es2015', 'react', 'stage-0']
+    presets        : ['es2015', 'react', 'stage-0'],
   },
   compiler_devtool         : 'source-map',
   compiler_hash_type       : 'hash',
@@ -43,13 +43,13 @@ const config = {
   compiler_stats           : {
     chunks : false,
     chunkModules : false,
-    colors : true
+    colors : true,
   },
   compiler_vendors : [
     'react',
     'react-redux',
     'react-router',
-    'redux'
+    'redux',
   ],
 
   // ----------------------------------
@@ -57,8 +57,8 @@ const config = {
   // ----------------------------------
   coverage_reporters : [
     { type : 'text-summary' },
-    { type : 'lcov', dir : 'coverage' }
-  ]
+    { type : 'lcov', dir : 'coverage' },
+  ],
 };
 
 /************************************************
@@ -76,14 +76,14 @@ Edit at Your Own Risk
 // N.B.: globals added here must _also_ be added to .eslintrc
 config.globals = {
   'process.env'  : {
-    'NODE_ENV' : JSON.stringify(config.env)
+    NODE_ENV : JSON.stringify(config.env),
   },
-  'NODE_ENV'     : config.env,
-  '__DEV__'      : config.env === 'development',
-  '__PROD__'     : config.env === 'production',
-  '__TEST__'     : config.env === 'test',
-  '__COVERAGE__' : !argv.watch && config.env === 'test',
-  '__BASENAME__' : JSON.stringify(process.env.BASENAME || '')
+  NODE_ENV     : config.env,
+  __DEV__      : config.env === 'development',
+  __PROD__     : config.env === 'production',
+  __TEST__     : config.env === 'test',
+  __COVERAGE__ : !argv.watch && config.env === 'test',
+  __BASENAME__ : JSON.stringify(process.env.BASENAME || ''),
 };
 
 // ------------------------------------
@@ -98,23 +98,23 @@ config.compiler_vendors = config.compiler_vendors
     debug(
       `Package "${dep}" was not found as an npm dependency in package.json; ` +
       `it won't be included in the webpack vendor bundle.
-       Consider removing it from \`compiler_vendors\` in ~/config/index.js`
+       Consider removing it from \`compiler_vendors\` in ~/config/index.js`,
     );
   });
 
 // ------------------------------------
 // Utilities
 // ------------------------------------
-function base () {
+function base() {
   const args = [config.path_base].concat([].slice.call(arguments));
-  return path.resolve.apply(path, args);
+  return path.resolve(...args);
 }
 
 config.paths = {
-  base   : base,
+  base,
   client : base.bind(null, config.dir_client),
   public : base.bind(null, config.dir_public),
-  dist   : base.bind(null, config.dir_dist)
+  dist   : base.bind(null, config.dir_dist),
 };
 
 // ========================================================
