@@ -10,15 +10,14 @@ dotenv.load();
 const User = require('../models/user');
 
 module.exports = function (passport) {
-
   // used to serialize the user for the session
-  passport.serializeUser(function (user, done) {
+  passport.serializeUser((user, done) => {
     done(null, user.id);
   });
 
    // used to deserialize the user
-  passport.deserializeUser(function (id, done) {
-    User.findById(id, function (err, user) {
+  passport.deserializeUser((id, done) => {
+    User.findById(id, (err, user) => {
       done(err, user);
     });
   });
@@ -26,29 +25,29 @@ module.exports = function (passport) {
   passport.use('local-signup', new LocalStrategy({
     usernameField: 'email',
     passwordField: 'password',
-    passReqToCallback : true
+    passReqToCallback: true,
   },
-  function (req, email, password, done) {
-    process.nextTick(function () {
-      User.findOne({ 'local.email' : email }, function (err, user) {
+  (req, email, password, done) => {
+    process.nextTick(() => {
+      User.findOne({ 'local.email': email }, (err, user) => {
         if (err) return done(err);
 
         if (user) {
           return done(err);
-        } else {
-          var newUser = new User();
+        }
+        const newUser = new User();
 
-          newUser.local.email = email;
-          newUser.local.password = newUser.generateHash(password);
-          newUser.local.firstName = req.body.firstName;
-          newUser.local.lastName = req.body.lastName;
+        newUser.local.email = email;
+        newUser.local.password = newUser.generateHash(password);
+        newUser.local.firstName = req.body.firstName;
+        newUser.local.lastName = req.body.lastName;
 
           // save the user
-          newUser.save(function (err) {
-            if (err) return done(err);
-            return done(null, newUser);
-          });
-        } // end else
+        newUser.save((err) => {
+          if (err) return done(err);
+          return done(null, newUser);
+        });
+         // end else
       });
     });
   }));
@@ -56,9 +55,9 @@ module.exports = function (passport) {
   passport.use('local-login', new LocalStrategy({
     usernameField: 'email',
     passwordField: 'password',
-    passReqToCallback : true
-  }, function (req, email, password, done) {
-    User.findOne({ 'local.email' : email }, function (err, user) {
+    passReqToCallback: true,
+  }, (req, email, password, done) => {
+    User.findOne({ 'local.email': email }, (err, user) => {
       if (err) return done(err);
 
       if (!user) return done(err);
@@ -72,30 +71,29 @@ module.exports = function (passport) {
   passport.use('google', new GoogleStrategy({
     clientID: process.env.GMAIL_ID,
     clientSecret: process.env.GMAIL_SECRET,
-    callbackURL: 'http://localhost:3000/auth/google/callback'
+    callbackURL: 'http://localhost:3000/auth/google/callback',
   },
-  function (token, refreshToken, profile, done) {
-    process.nextTick(function () {
-      User.findOne({ 'google.id': profile.id }, function (err, user) {
+  (token, refreshToken, profile, done) => {
+    process.nextTick(() => {
+      User.findOne({ 'google.id': profile.id }, (err, user) => {
         if (err) console.log(err);
 
         if (user) {
           // if a user is found, log them in
           return done(null, user);
-        } else {
-          var newUser = new User();
+        }
+        const newUser = new User();
 
-          newUser.google.id = profile.id;
-          newUser.google.token = token;
-          newUser.google.name = profile.displayName;
-          newUser.google.email = profile.emails[0].value; // pull the first email
+        newUser.google.id = profile.id;
+        newUser.google.token = token;
+        newUser.google.name = profile.displayName;
+        newUser.google.email = profile.emails[0].value; // pull the first email
 
           // save the user
-          newUser.save(function (err) {
-            if (err) throw err;
-            return done(null, newUser);
-          });
-        }
+        newUser.save((err) => {
+          if (err) throw err;
+          return done(null, newUser);
+        });
       });
     });
   }));
@@ -103,33 +101,32 @@ module.exports = function (passport) {
   passport.use(new GitHubStrategy({
     clientID: process.env.GITHUB_ID,
     clientSecret: process.env.GITHUB_SECRET,
-    callbackURL: 'http://localhost:3000/auth/github/callback'
+    callbackURL: 'http://localhost:3000/auth/github/callback',
   },
-  function (accessToken, refreshToken, profile, done) {
-    process.nextTick(function () {
-      User.findOne({ 'github.id': profile.id }, function (err, user) {
+  (accessToken, refreshToken, profile, done) => {
+    process.nextTick(() => {
+      User.findOne({ 'github.id': profile.id }, (err, user) => {
         if (err) console.log(err);
 
         if (user) {
           return done(null, user);
-        } else {
-          var newUser = new User();
+        }
+        const newUser = new User();
 
-          if (profile._json.email) {
-            newUser.github.email = profile._json.email;
-          }
-          newUser.github.id = profile._json.id;
-          newUser.github.name = profile._json.name;
-          newUser.github.token = accessToken;
+        if (profile._json.email) {
+          newUser.github.email = profile._json.email;
+        }
+        newUser.github.id = profile._json.id;
+        newUser.github.name = profile._json.name;
+        newUser.github.token = accessToken;
 
           // save the user
-          newUser.save(function (err) {
-            if (err) return done(err);
-            return done(null, newUser);
-          });
-        }
+        newUser.save((err) => {
+          if (err) return done(err);
+          return done(null, newUser);
+        });
       });
     });
-  }
+  },
 ));
 }; // end module
