@@ -2,13 +2,13 @@ import { connect } from 'react-redux';
 import { addOption, removeOption } from '../../modules/home';
 import Home from './Home';
 
-const mapStatetoProps = (state) => ({
-  options: state.home.options
+const mapStatetoProps = state => ({
+  options: state.home.options,
 });
 
 const mapDispatchToProps = {
   addOption,
-  removeOption
+  removeOption,
 };
 
 export default connect(mapStatetoProps, mapDispatchToProps)(Home);
