@@ -2,12 +2,12 @@ import { connect } from 'react-redux';
 import { setTabIndex } from '../../modules/tabmenu';
 import TabMenu from './TabMenu';
 
-const mapStatetoProps = (state) => ({
-  tabIndex: state.tabmenu.tabIndex
+const mapStatetoProps = state => ({
+  tabIndex: state.tabmenu.tabIndex,
 });
 
 const mapDispatchToProps = {
-  setTabIndex
+  setTabIndex,
 };
 
 export default connect(mapStatetoProps, mapDispatchToProps)(TabMenu);
