@@ -21,7 +21,7 @@ let render = () => {
 
   ReactDOM.render(
     <AppContainer store={store} routes={routes} />,
-    MOUNT_NODE
+    MOUNT_NODE,
   );
 };
 
@@ -51,7 +51,7 @@ if (__DEV__) {
       setImmediate(() => {
         ReactDOM.unmountComponentAtNode(MOUNT_NODE);
         render();
-      })
+      }),
     );
   }
 }
