@@ -7,17 +7,17 @@ export const REMOVE_OPTION = 'REMOVE_OPTION';
 // ------------------------------------
 // Actions
 // ------------------------------------
-export function addOption (index) {
+export function addOption(index) {
   return {
     type: ADD_OPTION,
-    index: index
+    index,
   };
 }
 
-export function removeOption (index) {
+export function removeOption(index) {
   return {
     type: REMOVE_OPTION,
-    index: index
+    index,
   };
 }
 
@@ -25,14 +25,14 @@ export function removeOption (index) {
 // Action Handlers
 // ------------------------------------
 const ACTION_HANDLERS = {
-  [ADD_OPTION] : (state, action) => ({
+  [ADD_OPTION]: (state, action) => ({
     ...state,
-    options: state.options.concat({ index: action.index + 1 })
+    options: state.options.concat({ index: action.index + 1 }),
   }),
-  [REMOVE_OPTION] : (state, action) => ({
+  [REMOVE_OPTION]: (state, action) => ({
     ...state,
-    options: [...state.options.slice(0, action.index), ...state.options.slice(action.index + 1)]
-  })
+    options: [...state.options.slice(0, action.index), ...state.options.slice(action.index + 1)],
+  }),
 };
 
 // ------------------------------------
@@ -41,11 +41,11 @@ const ACTION_HANDLERS = {
 const initialState = {
   options: [
     { index: 0 },
-    { index: 1 }
-  ]
+    { index: 1 },
+  ],
 };
 
-export default function navbarReducer (state = initialState, action) {
+export default function navbarReducer(state = initialState, action) {
   const handler = ACTION_HANDLERS[action.type];
 
   return handler ? handler(state, action) : state;
