@@ -19,17 +19,17 @@ describe('(Component) Header', () => {
   describe('Navigation links...', () => {
     it('Should render a Link to Home route', () => {
       expect(_wrapper.contains(
-        <IndexLink activeClassName='route--active' to='/'>
+        <IndexLink activeClassName="route--active" to="/">
           Home
-        </IndexLink>
+        </IndexLink>,
       )).to.be.true;
     });
 
     it('Should render a Link to Counter route', () => {
       expect(_wrapper.contains(
-        <Link activeClassName='route--active' to='/counter'>
+        <Link activeClassName="route--active" to="/counter">
           Counter
-        </Link>
+        </Link>,
       )).to.be.true;
     });
   });
