@@ -4,16 +4,18 @@ import { Counter } from 'routes/Counter/components/Counter';
 import { shallow } from 'enzyme';
 
 describe('(Component) Counter', () => {
-  let _props, _spies, _wrapper;
+  let _props,
+    _spies,
+    _wrapper;
 
   beforeEach(() => {
     _spies = {};
     _props = {
-      counter : 5,
+      counter: 5,
       ...bindActionCreators({
-        doubleAsync : (_spies.doubleAsync = sinon.spy()),
-        increment   : (_spies.increment = sinon.spy())
-      }, _spies.dispatch = sinon.spy())
+        doubleAsync: (_spies.doubleAsync = sinon.spy()),
+        increment: (_spies.increment = sinon.spy()),
+      }, _spies.dispatch = sinon.spy()),
     };
     _wrapper = shallow(<Counter {..._props} />);
   });
