@@ -1,5 +1,5 @@
 import {
-  default as createStore
+  default as createStore,
 } from 'store/createStore';
 
 describe('(Store) createStore', () => {
@@ -17,11 +17,11 @@ describe('(Store) createStore', () => {
   describe('(Location)', () => {
     it('store should be initialized with Location state', () => {
       const location = {
-        pathname : '/echo'
+        pathname: '/echo',
       };
       store.dispatch({
-        type    : 'LOCATION_CHANGE',
-        payload : location
+        type: 'LOCATION_CHANGE',
+        payload: location,
       });
       expect(store.getState().location).to.deep.equal(location);
     });
